requirements.txt
requests
requests_cache
bs4
ffmpy
pillow